+++ CherryCrisis/CherryEditor/src/core/main.cpp
@@ -101,7 +101,6 @@ int main()
         InputManager::GetInstance()->UpdateKeys();
         glfwPollEvents();
 
-
         ImGui_ImplOpenGL3_NewFrame();
         ImGui_ImplGlfw_NewFrame();
         ImGui::NewFrame();