+++ CherryCrisis/CherryEditor/src/panels/project_settings_displayer.cpp
@@ -123,17 +123,17 @@ void ProjectSettingsDisplayer::Input::Fill()
             if (success == 1)
             {
                 std::string notif = "Action Axes " + std::string(name) + " created.";
-                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Success, 1.0f);
+                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Success, 1.0f);
             }
             else if (success == 0)
             {
                 std::string notif = "Action Axes " + std::string(name) + " already exists.";
-                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Warning, 1.0f);
+                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Warning, 1.0f);
             }
             else if (success == -1)
             {
                 std::string notif = "Action Axes " + std::string(name) + " couldn't be created.";
-                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Error, 1.0f);
+                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Error, 1.0f);
             }
 
             memset(name, 0, IM_ARRAYSIZE(name));
@@ -168,17 +168,17 @@ void ProjectSettingsDisplayer::Input::Fill()
             if (success == 1)
             {
                 std::string notif = "Action Buttons " + std::string(name) + " created.";
-                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Success, 1.0f);
+                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Success, 1.0f);
             }
             else if (success == 0)
             {
                 std::string notif = "Action Buttons " + std::string(name) + " already exists.";
-                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Warning, 1.0f);
+                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Warning, 1.0f);
             }
             else if (success == -1)
             {
                 std::string notif = "Action Buttons " + std::string(name) + " couldn't be created.";
-                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Error, 1.0f);
+                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Error, 1.0f);
             }
 
             memset(name, 0, IM_ARRAYSIZE(name));
@@ -266,12 +266,12 @@ void ProjectSettingsDisplayer::Input::Fill()
                 if (success == 1)
                 {
                     std::string notif = "New axis added to Action.";
-                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Success, 1.0f);
+                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Success, 1.0f);
                 }
                 else if (success == -1)
                 {
                     std::string notif = "Couldn't add axis to Action.";
-                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Error, 1.0f);
+                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Error, 1.0f);
                 }
             }
             ImGui::TreePop();
@@ -316,7 +316,7 @@ void ProjectSettingsDisplayer::Input::Fill()
                             if (success == 0)
                             {
                                 std::string notif = "Key " + std::string(IM->GetKeyname(m_inputsIndex[j])) + " is already link to Action " + button.first;
-                                EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Warning, 1.0f);
+                                EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Warning, 1.0f);
                             }
                         }
 
@@ -343,12 +343,12 @@ void ProjectSettingsDisplayer::Input::Fill()
                 if (success == 1)
                 {
                     std::string notif = "New input added to Action.";
-                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Success, 1.0f);
+                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Success, 1.0f);
                 }
                 else if (success == -1)
                 {
                     std::string notif = "Couldn't add input to Action.";
-                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType_::ImGuiToastType_Error, 1.0f);
+                    EditorManager::SendNotification(notif.c_str(), ImGuiToastType::Error, 1.0f);
                 }
             }
             ImGui::TreePop();