+++ CherryCrisis/CherryEngine/include/engine.hpp
@@ -12,10 +12,8 @@ public:
 
 	bool isPlaying = false;
 
-	static void PlayFile(const char* fileNam);
 	void Launch();
 	void Stop() { isPlaying = false; }
-	void SoundInit();
 
 	void Tick();
 	void TickEngine();