+++ CherryCrisis/CherryEngine/include/renderer.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
-class ARenderer
+#include "cherry_macros.hpp"
+
+class CCENGINE_API ARenderer
 {
 public:
 	bool m_isVisible = true;