+++ CherryCrisis/CherryEngine/src/skybox_renderpass.cpp
@@ -92,7 +92,7 @@ void SkyboxRenderPass::Execute(const float x, const float y)
 	{
 		glBindVertexArray(gpuMesh->VAO);
 		glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, gpuMesh->EBO);
-		glDrawElements(GL_TRIANGLES, mesh->m_indices.size(), GL_UNSIGNED_INT, nullptr);
+		glDrawElements(GL_TRIANGLES, (GLsizei)mesh->m_indices.size(), GL_UNSIGNED_INT, nullptr);
 	}
 
 	if (Cubemap* skyTexture = m_skybox->m_cubemap.get())
@@ -114,6 +114,8 @@ int SkyboxRenderPass::Generate(CameraComponent* toGenerate)
 		return -1;
 
 	m_cameraComp = toGenerate;
+
+	return 1;
 }
 
 template <>